# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bracket Golf API - U.S. Amateur Bracket Challenge
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /groups:
    get:
      operationId: getGroups
      tags: [groups]
      summary: Get groups the current user belongs to
      responses:
        "200":
          description: User's groups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"
    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a new group
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/join:
    post:
      operationId: joinGroup
      tags: [groups]
      summary: Join a group by join code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinGroupInput"
      responses:
        "200":
          description: Joined group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{groupId}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get a group with its leaderboard
      parameters:
        - name: groupId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group with leaderboard
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupWithLeaderboard"

  # --- ADMIN ---
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
...
    LeaderboardEntry:
      type: object
      required: [rank, bracketId, bracketName, userId, userName, score, maxPossibleScore]
      properties:
        rank:
          type: number
        bracketId:
          type: string
        bracketName:
          type: string
        userId:
          type: string
        userName:
          type: string
        score:
          type: number
        maxPossibleScore:
          type: number
        championName:
          type: ["string", "null"]
...
    GroupInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        description:
          type: string
    JoinGroupInput:
      type: object
      required: [joinCode]
      properties:
        joinCode:
          type: string
    Group:
      type: object
      required: [id, tournamentId, name, joinCode, createdByUserId, memberCount, createdAt]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        joinCode:
          type: string
        createdByUserId:
          type: string
        memberCount:
          type: number
        createdAt:
          type: string
    GroupWithLeaderboard:
      type: object
      required: [id, tournamentId, name, joinCode, createdByUserId, memberCount, members, leaderboard, createdAt]
      properties:
        id:
          type: string
        tournamentId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        joinCode:
          type: string
        createdByUserId:
          type: string
        memberCount:
          type: number
        members:
          type: array
          items:
            $ref: "#/components/schemas/User"
        leaderboard:
          type: array
          items:
            $ref: "#/components/schemas/LeaderboardEntry"
        createdAt:
          type: string

    # --- ADMIN ---
...