# lib/api-zod/src/generated/types/joinGroupInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Bracket Golf API - U.S. Amateur Bracket Challenge
 * OpenAPI spec version: 0.1.0
 */

export interface JoinGroupInput {
  joinCode: string;
}